requirements.txt
panda
scikit-learn
regex
pyyaml
plotly
